requirements.txt
scikit-learn==0.24.2
matplotlib==3.3.3
telegram
airflow.operators.telegram_operator